+++ src/main/java/racingcar/MainController.java
@@ -1,6 +1,8 @@
 package racingcar;
 
 import java.util.List;
+import racingcar.view.InputView;
+import racingcar.view.OutputView;
 
 public class MainController {
